Release 2.0

+++ src/scripts/ona_service/hostname_resolver.py
@@ -16,49 +16,79 @@ from __future__ import print_function, unicode_literals
 # python builtins
 import json
 import logging
+import os
 import socket
+import subprocess
 from tempfile import NamedTemporaryFile
 from time import sleep
 
 # local
 from service import Service
 
+FORMAT = '%(asctime)s - %(name)s - %(levelname)s - %(message)s'
+logging.basicConfig(level=logging.INFO, format=FORMAT)
+
 DATA_TYPE = 'hostnames'
 UPDATE_INTERVAL_SECONDS = 600
 HOSTNAME_RESOLUTION_INTERVAL_SECONDS = 0.1
+ENV_HOSTNAME_DNS = 'OBSRVBL_HOSTNAME_DNS'
+ENV_HOSTNAME_NETBIOS = 'OBSRVBL_HOSTNAME_NETBIOS'
 
 
-def resolve_host_name(ip):
+def gethostbyaddr(ip):
     """
-    Resolve a single host name. Returns a string on success.
+    Resolve a single host name with gethostbyaddr. Returns a string on success.
     If resolution fails, returns None.
     """
     host = None
     try:
         host = socket.gethostbyaddr(ip)[0]
-    except socket.herror as e:
-        logging.info('Failed to resolve "{}": {}'.format(ip, e))
+    except socket.error:
+        pass
     return host
 
 
-def resolve_host_names(unresolved_ips):
+def nmblookup(ip, timeout_sec=1):
+    """
+    Resolve a single host name with nmblookup. Returns a string on success.
+    If resolution fails, returns None.
+    """
+    host = None
+    try:
+        output = subprocess.check_output(
+            ['timeout', '{}s'.format(timeout_sec), 'nmblookup', '-A', ip]
+        )
+    except subprocess.CalledProcessError:
+        pass
+    else:
+        for line in output.splitlines():
+            if '<GROUP>' in line:
+                continue
+            if ('<ACTIVE>' in line) and ('<00>' in line):
+                fields = line.strip().split()
+                host = fields[0].lower()
+                break
+
+    return host
+
+
+def resolve_host_names(unresolved_ips, resolvers):
     """
     for each ip, resolve its hostname. Don't overwhelm the router.
     Returns a dict of ip => hostname. Does not batch anything,
     this is a one-shot deal.
     """
-    resolved_ips = dict()
-    total = len(unresolved_ips)
-    processed = 0
-
+    ret = {}
     for ip in unresolved_ips:
-        resolved_ips[ip] = resolve_host_name(ip)
-        processed += 1
-        logging.info('Resolved "{}" to "{}" ({} of {}).'.format(
-            ip, resolved_ips[ip], processed, total))
+        name = None
+        for resolver in resolvers:
+            name = resolver(ip)
+            if name:
+                break
+        ret[ip] = name
         sleep(HOSTNAME_RESOLUTION_INTERVAL_SECONDS)
 
-    return resolved_ips
+    return ret
 
 
 class HostnameResolver(Service):
@@ -66,11 +96,19 @@ class HostnameResolver(Service):
     Routinely queries Observable Networks for interesting IPs, fetches their
     hostnames, and pushes the results to the ON cloud.
     """
+    resolver_map = {'gethostbyaddr': gethostbyaddr, 'nmblookup': nmblookup}
+
     def __init__(self, *args, **kwargs):
         kwargs.update({
             'poll_seconds': UPDATE_INTERVAL_SECONDS,
         })
         super(HostnameResolver, self).__init__(*args, **kwargs)
+
+        self.resolvers = []
+        if os.environ.get(ENV_HOSTNAME_NETBIOS, 'false') == 'true':
+            self.resolvers.append(nmblookup)
+        if os.environ.get(ENV_HOSTNAME_DNS, 'true') == 'true':
+            self.resolvers.append(gethostbyaddr)
 
     def _get_candidate_ips(self):
         try:
@@ -86,14 +124,18 @@ class HostnameResolver(Service):
             f.write(json.dumps(resolved))
             f.seek(0)
             path = self.api.send_file(DATA_TYPE, f.name, now, suffix='hosts')
-            data = {'path': path}
-            self.api.send_signal(DATA_TYPE, data)
+            if path is not None:
+                data = {'path': path}
+                self.api.send_signal(DATA_TYPE, data)
 
     def execute(self, now=None):
+        if not self.resolvers:
+            logging.error('No resolvers are set')
+
         ips = self._get_candidate_ips()
         if not ips:
             return
-        resolved = resolve_host_names(ips)
+        resolved = resolve_host_names(ips, self.resolvers)
         self._update_host_names(resolved, now)
